requirements.txt
pbixray==0.5.0
python-docx==1.1.2
